requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
groq
python-dotenv
pydantic
